Organize log-in form again
