+++ frontend/src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from "@/components/login-form"
+import { LoginForm } from "@/components/ui/login-form"
 
 export default function Page() {
   return (
